apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
httpx
google-generativeai
python-dotenv
supabase
slowapi
redis
celery
prometheus-client
structlog
